chore: Move client mutation types to separate file
