+++ application/ui/src/providers.tsx
@@ -10,7 +10,7 @@ import { ThemeProvider } from '@geti/ui/theme';
 import { QueryClientProvider } from '@tanstack/react-query';
 import { RouterProvider } from 'react-router';
 
-import { queryClient } from './query-client';
+import { queryClient } from './query-client/query-client';
 import { router } from './routes/router';
 
 export const Providers = () => {
