+++ application/ui/src/query-client/query-client.ts
@@ -3,39 +3,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { type paths } from '@geti-prompt/api';
-import { matchQuery, MutationCache, QueryClient, type Query } from '@tanstack/react-query';
-import type { HttpMethod } from 'openapi-typescript-helpers';
+import { matchQuery, MutationCache, Query, QueryClient } from '@tanstack/react-query';
 
-type OperationFor<Paths extends paths, P extends keyof Paths, Method extends HttpMethod> = Method extends keyof Paths[P]
-    ? Paths[P][Method]
-    : never;
-
-type PathParamsFor<Paths extends paths, P extends keyof Paths, Method extends HttpMethod> =
-    OperationFor<Paths, P, Method> extends { parameters: { path: infer PP } } ? PP : never;
-
-type MethodsForPath<Paths extends paths, P extends keyof Paths> = Extract<keyof Paths[P], HttpMethod>;
-
-export type QueryKey<Paths extends paths> = {
-    [P in keyof Paths]: {
-        [M in MethodsForPath<Paths, P>]: PathParamsFor<Paths, P, M> extends never
-            ? [M, P]
-            : [
-                  M,
-                  P,
-                  {
-                      params: {
-                          path: PathParamsFor<Paths, P, M>;
-                      };
-                  },
-              ];
-    }[MethodsForPath<Paths, P>];
-}[keyof Paths];
-
-type MutationMeta = {
-    invalidates?: QueryKey<paths>[];
-    awaits?: QueryKey<paths>[];
-};
+import { MutationMeta } from './query-client.interface';
 
 declare module '@tanstack/react-query' {
     interface Register {
